Addresses feedback points 1-7 and 8 from Anne van Kesteren: http://lists.w3.org/Archives/Public/public-pointer-events/2014JanMar/0069.html
Per WG action:
https://www.w3.org/2012/pointerevents/track/actions/72


> 1. PointerEvent constructor is not defined. You probably just want to reference http://dom.spec.whatwg.org/

At the end of Section 5.1, added a reference to your definition in DOM4.

> 2. The only place that maps event names (types) to their corresponding interface is non-normative.

added Section 5.2.1 that details the steps to fire a pointer event using the PointerEvent interface.  Then, in each of the places throughout the spec (mostly sections 5.2.3 - 5.2.12) where an even is fired, these steps are now referenced.

> 3. Instead of dispatch you want to talk about firing events, see http://dom.spec.whatwg.org/#concept-event-fire for the general idea and http://xhr.spec.whatwg.org/#concept-event-fire-progress for how to do it on an Event subclass. You will have define how to set all the properties of the class as well.
The new section on firing pointer events (5.2.1) now hooks into the process for firing/dispatch an event as you've defined in DOM4. Also replaced all instances of "dispatch" with "fire a pointer event named", except in the cases where the spec literally means to describe specifically the act of dispatching.

> 4. You cannot talk about "asynchronously" dispatch this day and age.
> You'll have to define how your standard interacts with the event loop:
> http://www.whatwg.org/specs/web-apps/current-work/#event-loop

integrated this with the event loop / task queue definitions from HTML5.

> 5. Why is pointercancel cancelable?
It shouldn't be (we discussed this on our last WG call too).  Fixed, along with pointerenter/leave which have no defined cancelable actions either.

> 6. A bunch of events are cancelable. In order to define that properly you need to define something like a) Fire an event /x/; b) If /x/'s canceled flag is unset, run Y. Otherwise it is unclear what needs to happen.
This has also been included in the new section on firing a pointer event.

> 8. It would be much better if pointer capture became part of the overall dispatching model rather than a separate set of steps you have to merge in your head with what came before (and hope it still makes sense!).
integrated this into the new section on firing pointer events, which should make this clearer.