Correct inverted pan direction definition to match examples.
When we talk of "panning down" we generally mean the viewport is moving down
relative to the content (increasing scrollTop values) which means the
content (and finger) is moving up relative to the screen.  I had this right
in the non-normative examples I gave, but inverted in the formal definition.
This CL fixes that and attempts to clarify the wording.