Simplification and correction of primary pointer section
Due to the addition of multi-mouse scenario, the explanation of what a
primary pointer is can be condensed/simplified, rather than being split
into "primary mouse"/"primary touch"/"primary pointer".

This change also corrects the previously erroneous description of when
touch/pen pointers become primary - as hovered pen (and presumably, if
supported by hardware, hovered touch) is also primary, it's not really
related to `pointermove` but rather to which active pointer for that
type came first.