Simplify language
avoiding the handwavy/convoluted "non-continuous zoom" - calling a spade a spade, and saying explicitly which types of gestures (involving multiple taps/touches with timings, such as double-tap to zoom) are affected by this. We feel that since we don't actually define these gestures, this is within the limits of what we can say in the spec...and it makes it much easier to understand.
further tiny tweaks/simplifications