Major refactoring: refer to "direct manipulation" rather than "touch" (#350)
as `touch-action` does not simply apply to touch, but any other situation where an input serves to directly pan/zoom (e.g. stylus on touchscreen), making this explicit. Before, it was just a non-normative note, but this now makes it normative that `touch-action` applies to all direct manipulation scenarios for panning/zooming.

(still leaves a bit of cognitive dissonance with the fact that the CSS property is called touch-action,

* replace "touch" references to "direct manipulation"
* add "direct manipulation" to glossary and link to it strategically
* simplify panning/scrolling to just panning, expand "direct manipulation" to specifically call out "direct manipulation for panning and/or zooming"