Reword lists for coalesced and predicted events
Rewrote them / returned them to the framing of "these are the steps that the user agent needs to follow". Clarified that this is in fact what the UA does (not something aimed at authors), and that these steps must be taken "when the ... method is called" (so any retargeting etc must happen at that point)
Also, changed `coalesced event list` / `predicted event list` to plural `coalesced events list` / `predicted events list`. the use of singular `event` was a bit awkward from the start.