Move definition point for PointerEvent (#406)
Currently, references to `{{PointerEvent}}` / `<a>PointerEvent</a>` link to/land on the sentence "The PointerEventInit dictionary is used by the PointerEvent interface's constructor..." which comes *after* the WebIDL and explanation prose for all the properties/methods. https://w3c.github.io/pointerevents/#dom-pointerevent
With this change, the `<dfn>` there is removed, and as a result, references to PointerEvent now link to (thanks to Respec) to the actual WebIDL definition itself, which seems far more logical. This is in line (by the look of it) with how other specs handle this (see for instance https://www.w3.org/TR/uievents/#mouseevent which links to the WebIDL)