Repoint `contextmenu` definition to UIEVENTS spec
The reference to the `contextmenu` event in 4.2.12 links to the HTML
spec, even though any non-editorial mentions of that event –but not the
definition in the event list– were removed (apparently by mistake) in
whatwg/html#2742. That event was subsequently added to the UI Events
spec in w3c/uievents#279, and now the definition in the HTML spec's
event list has been removed in whatwg/html#7506. This change updates
the reference to link to the UI Events spec.