Define a section for suppressing a pointer event stream.
The idea has been specified in multiple places of the spec through almost-identical text.  This PR consolidates all such text into a single section.

The key term used is "suppressing" because the previous term used ("canceling") was misleading because of an unrelated use ("event canceling" or prevent-defaulting).

Also normalized the spellings of "canceling" and "canceled" (to drop "ll"s).